unified_test/requirements.txt
Django
django-annoying
djangorestframework
django-model-utils
